@@ -12,7 +12,7 @@ license = "Apache-2.0"
 name = "pabi"
 readme = "README.md"
 repository = "https://github.com/kirillbobyrev/pabi"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 version = "0.0.0"
 # Reduce the package size by only including things necessary for building it.
 include = [
